split logic and fonted to diffrent files
