+++ src/components/NewPatientForm/FormSection.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+interface FormSectionProps {
+  title: string;
+  children: React.ReactNode;
+}
+
+export default function FormSection({ title, children }:FormSectionProps) {
+return (
+  <div className="patient-form__section">
+    <p className="patient-form__section-title">{title}</p>
+    <div className="patient-form__section-feilds">
+      {children}
+    </div>
+  </div>
+);
+}
