+++ src/components/NewPatientForm/SoldierDetailsSection.tsx
@@ -1,5 +1,8 @@
+import React from 'react';
 import EditField from "../EditField/EditField.tsx";
 import { FormErrors, PatientFormData } from "../../types/PatientForm.types.ts";
+import FormSection from './FormSection.tsx';
+
 
 interface Props {
   formData: PatientFormData;
@@ -13,7 +16,7 @@ interface Props {
   setIsDropoffTouched: (val: boolean) => void;
 }
 
-export default function SoldierDetailsSection({
+const SoldierDetailsSection: React.FC<Props> = ({
   formData,
   errors,
   onChange,
@@ -23,54 +26,51 @@ export default function SoldierDetailsSection({
   setIsPickupDisabled,
   setIsDropoffDisabled,
   setIsDropoffTouched,
-}: Props) {
-  return (
-    <div className="patient-form__section">
-      <p className="patient-form__section-title">פרטי החייל</p>
-      <div className="patient-form__section-feilds">
-        <EditField
-          title="שם מלא"
-          hint="הקלד שם החייל"
-          value={formData.fullName}
-          onTextChange={onChange("fullName")}
-          error={errors.fullName}
-          maxLength={20}
-        />
-        <EditField
-          title="טלפון"
-          type="tel"
-          hint="הקלד מספר"
-          value={formData.phone}
-          onTextChange={onChange("phone")}
-          error={errors.phone}
-        />
-        <EditField
-          title="תחנת איסוף"
-          showCheckbox
-          checked={!isPickupDisabled}
-          onCheckChange={(val) => setIsPickupDisabled(!val)}
-          options={["שגוב", "סיירים"]}
-          hint="בחר תחנה"
-          value={formData.pickupStation}
-          onTextChange={onPickupChange}
-          error={errors.pickupStation}
-        />
-        <EditField
-          title="תחנת פיזור"
-          showCheckbox
-          autoSize={{ minRows: 1, maxRows: 2 }}
-          checked={!isDropoffDisabled}
-          onCheckChange={(val) => setIsDropoffDisabled(!val)}
-          options={["שגוב", "סיירים"]}
-          hint="בחר תחנה"
-          value={isDropoffDisabled ? "" : formData.dropOffStation}
-          onTextChange={(val) => {
-            setIsDropoffTouched(true);
-            onChange("dropOffStation")(val);
-          }}
-          error={errors.dropOffStation}
-        />
-      </div>
-    </div>
-  );
-}
+}) => (
+  <FormSection title="פרטי החייל">
+    <EditField
+      title="שם מלא"
+      hint="הקלד שם החייל"
+      value={formData.fullName}
+      onTextChange={onChange("fullName")}
+      error={errors.fullName}
+      maxLength={20}
+    />
+    <EditField
+      title="טלפון"
+      type="tel"
+      hint="הקלד מספר"
+      value={formData.phone}
+      onTextChange={onChange("phone")}
+      error={errors.phone}
+    />
+    <EditField
+      title="תחנת איסוף"
+      showCheckbox
+      checked={!isPickupDisabled}
+      onCheckChange={(val) => setIsPickupDisabled(!val)}
+      options={["שגוב", "סיירים"]}
+      hint="בחר תחנה"
+      value={formData.pickupStation}
+      onTextChange={onPickupChange}
+      error={errors.pickupStation}
+    />
+    <EditField
+      title="תחנת פיזור"
+      showCheckbox
+      autoSize={{ minRows: 1, maxRows: 2 }}
+      checked={!isDropoffDisabled}
+      onCheckChange={(val) => setIsDropoffDisabled(!val)}
+      options={["שגוב", "סיירים"]}
+      hint="בחר תחנה"
+      value={isDropoffDisabled ? "" : formData.dropOffStation}
+      onTextChange={(val) => {
+        setIsDropoffTouched(true);
+        onChange("dropOffStation")(val);
+      }}
+      error={errors.dropOffStation}
+    />
+  </FormSection>
+);
+
+export default SoldierDetailsSection;
