+++ src/components/ShuttleTable/ShuttleTableHeader.tsx
@@ -1,0 +1,21 @@
+import { Flex, Segmented } from "antd";
+import { SizeType } from "antd/es/config-provider/SizeContext";
+import { useState } from "react";
+
+interface ShuttleTableHeaderProps {}
+export default function NewPatientFormContent({}: ShuttleTableHeaderProps) {
+  const [size, setSize] = useState<SizeType>("middle");
+  return (
+    <div className="shuttle-table-header">
+      <Flex gap="large" align="flex-start" vertical>
+        <Segmented
+        dir="ltr"
+          className="shuttle-table-header__segmented"
+          options={["הלוך", "חזור"]}
+          value={size}
+          onChange={(value) => setSize(value as SizeType)}
+        />
+      </Flex>
+    </div>
+  );
+}
