+++ src/scss/components/NewPatientForm/newpatientformcontent.scss
@@ -45,6 +45,8 @@
       column-gap: 3.2rem;
 
       width: 100%;
+
+      
     }
 
     &-title {
